requirements.txt
torch>=2.2.0
transformers>=4.39.0
datasets>=2.18.0
evaluate>=0.4.1
scikit-learn>=1.4.0
pandas>=2.2.0
numpy>=1.26.0
typer>=0.12.3
rich>=13.7.1
shap>=0.45.0
raid-bench>=0.3.0
